requirements.txt
-i https://pypi.org/simple

blinker==1.6.2 ; python_version >= '3.7'

click==8.1.3 ; python_version >= '3.7'

colorama==0.4.6 ; platform_system == 'Windows'

dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'

flask==2.3.2

itsdangerous==2.1.2 ; python_version >= '3.7'

jinja2==3.1.2 ; python_version >= '3.7'

markupsafe==2.1.2 ; python_version >= '3.7'

pymongo[srv]==4.3.3

python-dotenv==1.0.0

werkzeug==2.3.4 ; python_version >= '3.8'

gunicorn


